Optimize set date (add year)

+++ main/elster.h
@@ -96,12 +96,12 @@ void SetValueType(char * Val, uint8_t Type, uint16_t Value);
 void SetDoubleType(char * Val, unsigned char Type, double Value);
 // bool FormElsterTable(const KCanFrame & Frame, char * str);
 const char * ElsterTypeToName(ElsterValueType Type);
-uint16_t TranslateString(const char * str, uint8_t elster_type);
+uint32_t TranslateString(const char * str, uint8_t elster_type);
 
 ElsterPacketReceive ElsterRawToReceivePacket(uint16_t sender, uint8_t length, uint8_t const * const data);
 
 void ElsterPrepareSendPacket(uint8_t length, uint8_t * const data, ElsterPacketSend packet);
-void ElsterSetValueDefault(uint8_t length, uint8_t * const data, uint16_t value);
+void ElsterSetValueDefault(uint8_t length, uint8_t * const data, uint32_t value);
 void ElsterSetValueBool(uint8_t length, uint8_t * const data, bool value);
 
 #endif
